Crate disk

Structs

PartitionInfo
Information about a partition to create.

QcowFile
Represents a qcow2 file. This is a sparse file format maintained by the qemu project. Full documentation of the format can be found in the qemu repository.

SingleFileDisk
A disk backed by a single file that implements AsyncDisk for access.

Enums

CompositeError

Error

GptError

ImagePartitionType
The type of partition.

ImageType
The variants of image files on the host that can be used as virtual disks.

Constants

MAX_NESTING_DEPTH
Nesting depth limit for disk formats that can open other disk files.

QCOW_MAGIC

Traits

AsyncDisk
An asynchronously accessible disk.

DiskFile
The prerequisites necessary to support a block device.

DiskGetLen
A trait for getting the length of a disk image or raw block device.

ToAsyncDisk
A DiskFile that can be converted for asychronous access.

Functions

async_ok
Check if the image file type can be used for async disk access.

convert
Copy the contents of a disk image in src_file into dst_file. The type of src_file is automatically detected, and the output file type is determined by dst_type.

create_async_disk_file
Inspect the image file type and create an appropriate disk file to match it.

create_composite_disk
Create a new composite disk image containing the given partitions, and write it out to the given files.

create_disk_file
Inspect the image file type and create an appropriate disk file to match it.

create_zero_filler
Create a zero filler file which can be used to fill the gaps between partition files. The filler is sized to be big enough to fill the gaps. (1 << PARTITION_SIZE_SHIFT)

detect_image_type
Detect the type of an image file by checking for a valid header of the supported formats.

Type Definitions

Result