Struct disk::QcowFile

pub struct QcowFile {
    raw_file: QcowRawFile,
    header: QcowHeader,
    l1_table: VecCache<u64>,
    l2_entries: u64,
    l2_cache: CacheMap<VecCache<u64>>,
    refcounts: RefCount,
    current_offset: u64,
    unref_clusters: Vec<u64>,
    avail_clusters: Vec<u64>,
    backing_file: Option<Box<dyn DiskFile>>,
}

Represents a qcow2 file. This is a sparse file format maintained by the qemu project. Full documentation of the format can be found in the qemu repository.

Example

    let mut q = QcowFile::from(file, disk::MAX_NESTING_DEPTH).expect("Can't open qcow file");
    let mut buf = [0u8; 12];
    let mut vslice = VolatileSlice::new(&mut buf);
    q.read_at_volatile(vslice, 10)?;

Fields
raw_file: QcowRawFileheader: QcowHeaderl1_table: VecCache<u64>l2_entries: u64l2_cache: CacheMap<VecCache<u64>>refcounts: RefCountcurrent_offset: u64unref_clusters: Vec<u64>avail_clusters: Vec<u64>backing_file: Option<Box<dyn DiskFile>>
Implementations

impl QcowFile

pub fn from(file: File, max_nesting_depth: u32) -> Result<QcowFile, Error>

Creates a QcowFile from file. File must be a valid qcow2 image.

pub fn new(file: File, virtual_size: u64) -> Result<QcowFile, Error>

Creates a new QcowFile at the given path.

pub fn new_from_backing(
file: File,
backing_file_name: &str,
backing_file_max_nesting_depth: u32
) -> Result<QcowFile, Error>

Creates a new QcowFile at the given path.

fn new_from_header(
file: File,
header: QcowHeader,
max_nesting_depth: u32
) -> Result<QcowFile, Error>

pub fn set_backing_file(&mut self, backing: Option<Box<dyn DiskFile>>)

pub fn first_zero_refcount(&mut self) -> Result<Option<u64>, Error>

Returns the first cluster in the file with a 0 refcount. Used for testing.

fn find_avail_clusters(&mut self) -> Result<(), Error>

fn rebuild_refcounts(
raw_file: &mut QcowRawFile,
header: QcowHeader
) -> Result<(), Error>

Rebuild the reference count tables.

fn limit_range_file(&self, address: u64, count: usize) -> usize

fn limit_range_cluster(&self, address: u64, count: usize) -> usize

fn virtual_size(&self) -> u64

fn l1_address_offset(&self, address: u64) -> u64

fn l1_table_index(&self, address: u64) -> u64

fn l2_table_index(&self, address: u64) -> u64

fn file_offset_read(&mut self, address: u64) -> Result<Option<u64>>

fn file_offset_write(&mut self, address: u64) -> Result<u64>

fn update_cluster_addr(
&mut self,
l1_index: usize,
l2_index: usize,
cluster_addr: u64,
set_refcounts: &mut Vec<(u64, u16)>
) -> Result<()>

fn get_new_cluster(&mut self, initial_data: Option<Vec<u8>>) -> Result<u64>

fn append_data_cluster(&mut self, initial_data: Option<Vec<u8>>) -> Result<u64>

fn deallocate_cluster(&mut self, address: u64) -> Result<()>

fn zero_bytes(&mut self, address: u64, length: usize) -> Result<()>

fn read_l2_cluster(
raw_file: &mut QcowRawFile,
cluster_addr: u64
) -> Result<Vec<u64>>

fn set_cluster_refcount(
&mut self,
address: u64,
refcount: u16
) -> Result<Vec<u64>>

fn sync_caches(&mut self) -> Result<()>

fn read_cb<F>(&mut self, address: u64, count: usize, cb: F) -> Result<usize> where
F: FnMut(Option<&mut dyn DiskFile>, usize, u64, usize) -> Result<()>,

fn write_cb<F>(&mut self, address: u64, count: usize, cb: F) -> Result<usize> where
F: FnMut(&mut File, usize, u64, usize) -> Result<()>,