Module disk::qcow

Modules

qcow_raw_file

refcount

vec_cache

Structs

QcowFile

Represents a qcow2 file. This is a sparse file format maintained by the qemu project. Full documentation of the format can be found in the qemu repository.

QcowHeader

Contains the information from the header of a qcow file.

Enums

Error

Constants

CLUSTER_USED_FLAG

COMPATIBLE_FEATURES_LAZY_REFCOUNTS

COMPRESSED_FLAG

DEFAULT_CLUSTER_BITS

DEFAULT_REFCOUNT_ORDER

L1_TABLE_OFFSET_MASK

L2_TABLE_OFFSET_MASK

MAX_BACKING_FILE_SIZE

MAX_CLUSTER_BITS

MAX_QCOW_FILE_SIZE

MAX_RAM_POINTER_TABLE_SIZE

MIN_CLUSTER_BITS

QCOW_EMPTY_HEADER_EXTENSION_SIZE

QCOW_MAGIC

V3_BARE_HEADER_SIZE

Functions

div_round_up_u32

div_round_up_u64

max_refcount_clusters

offset_is_cluster_boundary

read_u16_from_file

read_u32_from_file

read_u64_from_file

Type Definitions

Result