Trait disk::AsyncDisk

pub trait AsyncDisk: DiskGetLen + FileSetLen + FileAllocate {
    fn into_inner(self: Box<Self>) -> Box<dyn DiskFile>;

    fn fsync<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<()>> + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn read_to_mem<'a, 'async_trait>(
        &'a self, 
        file_offset: u64, 
        mem: Arc<dyn BackingMemory + Send + Sync>, 
        mem_offsets: &'a [MemRegion]
    ) -> Pin<Box<dyn Future<Output = Result<usize>> + 'async_trait>>
    where
        'a: 'async_trait,
        Self: 'async_trait;

    fn write_from_mem<'a, 'async_trait>(
        &'a self, 
        file_offset: u64, 
        mem: Arc<dyn BackingMemory + Send + Sync>, 
        mem_offsets: &'a [MemRegion]
    ) -> Pin<Box<dyn Future<Output = Result<usize>> + 'async_trait>>
    where
        'a: 'async_trait,
        Self: 'async_trait;

    fn punch_hole<'life0, 'async_trait>(
        &'life0 self, 
        file_offset: u64, 
        length: u64
    ) -> Pin<Box<dyn Future<Output = Result<()>> + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn write_zeroes_at<'life0, 'async_trait>(
        &'life0 self, 
        file_offset: u64, 
        length: u64
    ) -> Pin<Box<dyn Future<Output = Result<()>> + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

An asynchronously accessible disk.

Required methods

fn into_inner(self: Box<Self>) -> Box<dyn DiskFile>

Returns the inner file consuming self.

fn fsync<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<()>> + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Asynchronously fsyncs any completed operations to the disk.

fn read_to_mem<'a, 'async_trait>(
&'a self,
file_offset: u64,
mem: Arc<dyn BackingMemory + Send + Sync>,
mem_offsets: &'a [MemRegion]
) -> Pin<Box<dyn Future<Output = Result<usize>> + 'async_trait>> where
'a: 'async_trait,
Self: 'async_trait,

Reads from the file at ‘file_offset’ in to memory mem at mem_offsets. mem_offsets is similar to an iovec except relative to the start of mem.

fn write_from_mem<'a, 'async_trait>(
&'a self,
file_offset: u64,
mem: Arc<dyn BackingMemory + Send + Sync>,
mem_offsets: &'a [MemRegion]
) -> Pin<Box<dyn Future<Output = Result<usize>> + 'async_trait>> where
'a: 'async_trait,
Self: 'async_trait,

Writes to the file at ‘file_offset’ from memory mem at mem_offsets.

fn punch_hole<'life0, 'async_trait>(
&'life0 self,
file_offset: u64,
length: u64
) -> Pin<Box<dyn Future<Output = Result<()>> + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Replaces a range of bytes with a hole.

fn write_zeroes_at<'life0, 'async_trait>(
&'life0 self,
file_offset: u64,
length: u64
) -> Pin<Box<dyn Future<Output = Result<()>> + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Writes up to length bytes of zeroes to the stream, returning how many bytes were written.